Efectos y fix bugs

+++ client/includes/client_protocol.h
@@ -77,7 +77,7 @@ public:
     void sendInteract(uint32_t npc_id, uint8_t action) const;
     void sendTakeItem() const;
     void sendBuyItem(uint32_t npc_id, uint8_t item_id, uint16_t quantity) const;
-    void sendSellItem(uint32_t npc_id, uint16_t item_id, uint16_t quantity) const;
+    void sendSellItem(uint32_t npc_id, uint8_t item_id, uint16_t quantity) const;
     void sendDisconnect() const;
     void sendListItems(Id npc_id);
     void sendDepositItem(Id npc_id, Id item_id);
