+++ client/src/commands/command_sell_item.cpp
@@ -1,6 +1,6 @@
 #include "client/includes/commands/command_sell_item.h"
 
-SellItemCommandClient::SellItemCommandClient(uint32_t npc_id, uint16_t item_id, uint16_t quantity):
+SellItemCommandClient::SellItemCommandClient(uint32_t npc_id, uint8_t item_id, uint16_t quantity):
         npc_id(npc_id), item_id(item_id), quantity(quantity) {}
 
 void SellItemCommandClient::execute(ClientProtocol& protocol) const {
